Make `Splotch` top-level and move icons into `icons` dir
