+++ frontend/src/lib/components/Cards/InsightCard/InsightCard.tsx
@@ -25,7 +25,7 @@ import {
     InsightModel,
     InsightType,
 } from '~/types'
-import { Splotch, SplotchColor } from 'lib/lemon-ui/icons/Splotch'
+import { Splotch, SplotchColor } from 'lib/lemon-ui/Splotch'
 import { LemonButton, LemonButtonWithPopup } from 'lib/lemon-ui/LemonButton'
 import { LemonDivider } from 'lib/lemon-ui/LemonDivider'
 import { Link } from 'lib/lemon-ui/Link'
