+++ frontend/src/lib/lemon-ui/Splotch/Splotch.scss
