+++ frontend/src/lib/lemon-ui/Splotch/index.ts
@@ -1,0 +1,1 @@
+export * from './Splotch'
