+++ frontend/src/lib/lemon-ui/icons/icons.scss
