+++ frontend/src/lib/lemon-ui/icons/icons.stories.tsx
@@ -1,10 +1,11 @@
 import * as React from 'react'
-import * as icons from 'lib/lemon-ui/icons'
+import * as icons from './icons'
 import { Meta } from '@storybook/react'
-import { LemonTable } from './LemonTable'
-import { IconGauge, IconWithCount } from 'lib/lemon-ui/icons'
-import { LemonCheckbox } from './LemonCheckbox'
-import { LemonButton } from './LemonButton'
+import { LemonTable } from 'lib/lemon-ui/LemonTable'
+import { LemonCheckbox } from 'lib/lemon-ui/LemonCheckbox'
+import { LemonButton } from 'lib/lemon-ui/LemonButton'
+
+const { IconGauge, IconWithCount } = icons
 
 interface IconDefinition {
     name: string
@@ -93,13 +94,7 @@ export function Library(): JSX.Element {
 
 export function IconWithCountBubble(): JSX.Element {
     return (
-        <span
-            style={{
-                display: 'inline-flex',
-                fontSize: '1.5rem',
-                border: '1px solid var(--primary)',
-            }}
-        >
+        <span className="inline-flex text-2xl border border-primary">
             <IconWithCount count={7}>
                 <IconGauge />
             </IconWithCount>
@@ -109,13 +104,7 @@ export function IconWithCountBubble(): JSX.Element {
 
 export function IconWithCountHidingZero(): JSX.Element {
     return (
-        <span
-            style={{
-                display: 'inline-flex',
-                fontSize: '1.5rem',
-                border: '1px solid var(--primary)',
-            }}
-        >
+        <span className="inline-flex text-2xl border border-primary">
             <IconWithCount count={0} showZero={false}>
                 <IconGauge />
             </IconWithCount>
@@ -125,13 +114,7 @@ export function IconWithCountHidingZero(): JSX.Element {
 
 export function IconWithCountShowingZero(): JSX.Element {
     return (
-        <span
-            style={{
-                display: 'inline-flex',
-                fontSize: '1.5rem',
-                border: '1px solid var(--primary)',
-            }}
-        >
+        <span className="inline-flex text-2xl border border-primary">
             <IconWithCount count={0} showZero={true}>
                 <IconGauge />
             </IconWithCount>
@@ -141,13 +124,7 @@ export function IconWithCountShowingZero(): JSX.Element {
 
 export function IconWithCountOverflowing(): JSX.Element {
     return (
-        <span
-            style={{
-                display: 'inline-flex',
-                fontSize: '1.5rem',
-                border: '1px solid var(--primary)',
-            }}
-        >
+        <span className="inline-flex text-2xl border border-primary">
             <IconWithCount count={11} showZero={true}>
                 <IconGauge />
             </IconWithCount>
